Refactors ClientIdInterface::saveKeyToFile
The original saveKeyToFile function didn't take into account the pair
may already exist in the file whether we might be adding a duplicate key
or updating an ID's key.  This should resolve that issue by essentially
rewriting the file each time to saveKeyToFile.

This does add some overhead to the function, but ideally the function
isn't being called often enough to make a noticeable difference.  There
may in fact be a better method but this refactor at least makes it
behave more appropriately.