Refactor ClientIdInterface::getDelimPos
getDelimPos now searches from the end of the string.  This should
be safer as it eliminates any sanitation issues if a client id were
to contain a delim char.  A key is purposely constructed to not contain
any delim characters

A potential optimization could be to simply calculate the end of a
id/key string and subtract the length of a key to end up exactly where
the delimiter should be.  The speed in which the file is rendered
however does not warrant this optimization.