QTextStream issues
Apparently QTextStream requires a device to operate on.  Instead a
simple QString is used.  This could potentially be an issue if the
number of client ids were to grow very large as the entire file is
loaded into memory.  However, the number of client ids would need to be
extremely large and is likely outside the bounds of expected number
clients.

There is also apparently an issue in generating the keys.  Instead of
using the entire string of available characters, only the first four
are used.  Further investigation necessary.