Challenge Verification/Socket Issues
Some simple status outputs are added however they still need to be
connected to the main UI in some way.  Some color coding would be
marvelous.

Made some small changes to the socket read/write calls as they were not
getting pushed out properly and instead were just sitting the buffer.
Those issues should be resolved.  However there seems to be a major
issue in the challenge verification system.  Challenge responses that should
match are not for some reason.  Debugging needed.

Another significant issue that has cropped up is in the server
application, once a client has connected at some point, the socket
thread does not seem to be terminating correctly which in turn is
causing the application to flat out crash upon attempting to close.

Some good news is that things are quickly approaching a near usable
state.  There are still some significant features still to be added and
I have a feeling the way mouse moves are sent will need a lot of
tweaking.  Progress!