Removes Id Len
Commented out the portions of code which gave a limitation to how long
a clients ID could be (8).  It came to pass that the length of the ID
was needed to be passed in the socket message which is contained to a
byte.  So technically there is a limit at 255 characters so I left the
code in there for the time being and instead just commented out the
check.