Changes to Key Chars and Bug fix
I decided to remove the letters 'i','I','l','L', and 'O' since it can
be difficult to interpret what exactly is what in the given context so
this made life easier.

There was also an 'off-by-one' error in parsing the id/key file where
the new line character was being included in the key which led to some
issues.