Socket issue fixes
There was an issue with the if/else statements in regards to figuring
out if the socket had closed or if there was an error which was part of
the issure of the program throwing an obscure error about thread events
when you closed the server.  It is only partially fixed at the moment
but I have a good idea of what needs to be changed in order to fix it.

Also parsing the data input for the length of the id string became
necessary so there is now a byte after the reqest message for the
length of the id that follows.  The id is required to pull up the
correct key for the given client id.