Client socket changes
Client socket now uses a more simplified 'BufferedOutputStream' instead
of the 'PrintWriter'  which is more condusive to binary writing;
specifically the hash check.  It also became more apparent that sending mouse
data in a binary format is more intuitive (and more effecient).

I hadn't noticed before the when shoving a byte array into the
PrintWriter object it was simply calling toString() on the byte array
which in turn reported the address at which the byte array was sitting
and not the contents of the array; woops, should've known better.