Still debugging challenge issues
Some progress was made in the vein of trying to make the challenge
verification work.  The incompatibilities between Java and C++ are
becoming more and more apparent.

Also fixed a bunch of code that made me facepalm.  Couldn't really
explain what I was thinking when I wrote it.

Next step in attempting to fix the challenge verification should be to
simply try a shorter challenge string, that should start pulling out
any final bugs.  Struggled for a while in figuring how to debug from
here but I think that will be the key.  It's tricky swapping between
QtCreator and Eclipse debugging environments and fairly painful
debugging them syncronously.  It's definitely easier to approach one at
time but feels less fruitful.