Client Verification finally working
Finally solves the issue with client verification not working
correctly.  There is still much that can be improved but at the very
least it's working correctly at this juncture.

Next item should be fixing the socket threading issue that is
preventing the server to close down correctly with a client connected.